feat(phase-4): preopen akua stdlib into worker at /akua-stdlib
The worker was cfg-gating out the \`akua\` ExternalPkg on wasm32
because \`akua_core::stdlib::stdlib_root()\` materializes the stdlib
under \`\$TMPDIR\` and \`std::env::temp_dir()\` is an unconditional
panic on wasip1. So Packages that did \`import akua.helm\` couldn't
run through the sandbox — only through the native path.

Flip it: the host materializes the stdlib (it can; temp_dir works
on host) and preopens that dir into every worker invocation at the
well-known guest path \`/akua-stdlib\`. The worker unconditionally
registers \`akua\` as an ExternalPkg pointing at that guest path, so
the same \`import akua.helm\` source resolves on both sides. No
protocol change — the preopen is implicit.

Also fixes a latent overflow in the engine-plugin Session init: we
were calling \`set_epoch_deadline(u64::MAX)\` to opt plugin Stores
out of the wall-clock cap, but wasmtime internally does
\`current_epoch + delta\`, so once the ticker advanced past zero the
add wrapped. Drop to \`i64::MAX\` — at the 100ms ticker rate that's
~29 billion years, effectively infinite without the overflow.

+++ crates/engine-host-wasm/src/lib.rs
@@ -171,9 +171,14 @@ impl Session {
         // Shared Engine has `epoch_interruption` enabled (for the
         // render worker's wall-clock cap). Engine plugins (helm,
         // kustomize) don't want a deadline — the host-Rust caller
-        // above us owns their whole-call timeouts. Set the highest
-        // deadline so the epoch-ticker never trips their Store.
-        store.set_epoch_deadline(u64::MAX);
+        // above us owns their whole-call timeouts. Pick a ceiling
+        // high enough that the ticker never trips their Store: the
+        // `set_epoch_deadline(delta)` API internally does
+        // `current_epoch + delta`, so `u64::MAX` overflows once the
+        // ticker has advanced past zero. `i64::MAX` (2^63-1) at the
+        // 100ms ticker rate is ~29 billion years — comfortably
+        // effectively-infinite.
+        store.set_epoch_deadline(i64::MAX as u64);
         let mut linker: Linker<WasiP1Ctx> = Linker::new(&engine);
         p1::add_to_linker_sync(&mut linker, |s: &mut WasiP1Ctx| s).map_err(wasm_err)?;
